+++ src/commands.ts
@@ -4,8 +4,10 @@
 
 import { IOpsgenieCommandDesc } from "./lib/opsgenie-command";
 import alertsCommand from "./commands/alerts";
+import integrationsCommand from "./commands/integrations";
 
 export const commands: IOpsgenieCommandDesc[] = [
     alertsCommand,
+    integrationsCommand,
 ];